requirements.txt
Flask
Gunicorn
psycopg2-binary
WTForms
Flask-WTF